edit APIs for Showdown, Shutdown and Turnaround

+++ apps/java/services/case-engine-rest-api/src/main/java/com/wks/caseengine/dto/product/ShutDownPlanDTO.java
@@ -12,7 +12,7 @@ public class ShutDownPlanDTO {
     private Date maintStartDateTime;
     private Date maintEndDateTime;
 	// Change from Integer to Long
-    private Long durationInMins; 
+    private Integer durationInMins; 
     
 	//FOR ID : pmt.Id
 	private UUID maintenanceId;
